src/com/zjxjwxk/leetcode/_0033_Search_in_Rotated_Sorted_Array/Solution.java
package com.zjxjwxk.leetcode._0033_Search_in_Rotated_Sorted_Array;

/**
 * 二分查找
 * @author Xinkang Wu
 * @date 2021/4/7 22:53
 */
public class Solution {

    public int search(int[] nums, int target) {
        int left = 0, right = nums.length - 1, mid;
        while (left <= right) {
            mid = left + ((right - left) >> 1);
            if (nums[mid] == target) {
                return mid;
            } else {
                if (nums[left] <= nums[mid]) {
                    if (target >= nums[left] && target < nums[mid]) {
                        right = mid - 1;
                    } else {
                        left = mid + 1;
                    }
                } else {
                    if (target > nums[mid] && target <= nums[right]) {
                        left = mid + 1;
                    } else {
                        right = mid - 1;
                    }
                }
            }
        }
        return -1;
    }
}
